backend/requirements.txt
# Creator OS - Backend Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database - usar versión estable
supabase==2.3.0

# HTTP Client
httpx>=0.24.0,<0.26

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Text Processing
nltk>=3.8.0
scikit-learn>=1.4.0

# Utilities
python-dateutil>=2.8.0

# Production
gunicorn>=21.0.0